Updated price, account, order and cart

See merge request Omnijar/Naamio/kauppa!76

+++ Sources/KauppaTax/Service/Validation.swift
@@ -4,9 +4,9 @@ import KauppaTaxModel
 /// Configuration object for the tax service.
 struct TaxConfiguration {
     /// Minimum allowed value for tax rate in percentage.
-    static var minTaxRatePercent: Double = 0.0
+    static var minTaxRatePercent: Float = 0.0
     /// Maximum allowed value for tax rate in percentage.
-    static var maxTaxRatePercent: Double = 1000.0   // disallow rates > 1000%
+    static var maxTaxRatePercent: Float = 1000.0    // disallow rates > 1000%
 }
 
 extension TaxRate {
